some queue functions & rearranging stuff
