+++ src/beefcommands/utilities/music_player/queue.py
@@ -1,23 +1,36 @@
 import discord
+import random
 
 queue = []
 
-async def queue_stack(self, interaction: discord.Interaction):
+async def queue_stack(self, interaction: discord.Interaction, tracks: list):
     await interaction.response.send_message("not implemented yet sorry :(")
-    #add add the link to the top of the queue
+    queue.insert(0, tracks[0])
     
-async def queue_push(self, interaction: discord.Interaction):
+async def queue_push(self, interaction: discord.Interaction, tracks: list):
     await interaction.response.send_message("not implemented yet sorry :(")
-    #add the link to the back of the queue
+    for track in tracks:
+        queue.append(track)
     
 async def queue_pop(self, interaction: discord.Interaction):
     await interaction.response.send_message("not implemented yet sorry :(")
-    #remove the first item in the queue
+    queue.pop(0)
     
 async def clear(self, interaction: discord.Interaction):
     await interaction.response.send_message("not implemented yet sorry :(")
-    #clear the queue
+    queue.clear()
     
 async def shuffle(self, interaction: discord.Interaction):
     await interaction.response.send_message("not implemented yet sorry :(")
-    #randomises the queue
+    random.shuffle
+    
+async def queue_list(self, interaction: discord.Interaction):
+    for track in queue:
+        print(track)
+    #generate an embed listing the first 10 songs in the queue,
+    # is there a way to dynamically add pages?
+    
+async def queue_embed(self, interaction: discord.Interaction):
+    pass
+#generate a queue embed
+# "xyz added abc to the back/front queue"
